requirements.txt
mysql-connector-python
requests
tqdm
python-dotenv